requirements.txt
fastapi
uvicorn
jinja2
pyyaml
markdown
python-multipart
groq
python-dotenv 